refactor: remove ElmSelect and related files, rename ElmSelect to ElmSelectSlot
