--- packages/qwik/src/components/form/elm-select-slot.module.css
@@ -1,116 +1,0 @@
-.wrapper {
-  --highlight-color: #bfa056;
-
-  user-select: none;
-  position: relative;
-  box-sizing: border-box;
-  padding: 0.25rem;
-  border-radius: 0.25rem;
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 0.25rem;
-
-  border-style: solid;
-  border-width: 1px;
-  border-color: transparent;
-
-  transition:
-    border-color 200ms,
-    background-color 200ms;
-
-  background-color: rgba(255, 255, 255, 0.8);
-  box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.15);
-  cursor: pointer;
-
-  &.active {
-    border-color: var(--highlight-color);
-  }
-
-  [data-theme="dark"] & {
-    background-color: rgba(255, 255, 255, 0.15);
-    box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.75);
-  }
-
-  .label {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.125rem;
-    margin: 0;
-    padding: 0;
-    font-size: 0.75rem;
-    line-height: 0.75rem;
-    height: 0.75rem;
-    transition: color 200ms;
-    user-select: none;
-    color: gray;
-
-    &.label-active {
-      color: var(--highlight-color);
-    }
-  }
-
-  .body {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 0.25rem;
-    height: 1.25rem;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    .selected-option {
-      overflow: hidden;
-    }
-
-    .fallback {
-      opacity: 0.6;
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-      gap: 0.5rem;
-    }
-
-    .pulldown {
-      position: absolute;
-      top: calc(100% + 0.25rem);
-      left: 0;
-      z-index: 100;
-
-      box-sizing: border-box;
-      width: max-content;
-      padding: 0;
-      border-radius: 0.25rem;
-
-      background-color: rgba(255, 255, 255, 0.95);
-      box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.15);
-
-      [data-theme="dark"] & {
-        background-color: rgba(51, 51, 51, 0.95);
-        box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.75);
-      }
-
-      .option {
-        overflow: hidden;
-        width: 100%;
-        padding: 0.5rem;
-        box-sizing: border-box;
-        border-radius: 0.25rem;
-        display: flex;
-        justify-content: flex-start;
-        align-items: center;
-        gap: 0.25rem;
-        transition: background-color 100ms;
-        cursor: pointer;
-
-        &:hover {
-          background-color: rgba(190, 194, 202, 0.3);
-          [data-theme="dark"] & {
-            background-color: rgba(190, 194, 202, 0.2);
-          }
-        }
-      }
-    }
-  }
-}
