+++ packages/qwik/src/components/form/elm-select.module.css
@@ -1,13 +1,15 @@
 .wrapper {
+  --highlight-color: #bfa056;
+
   user-select: none;
   position: relative;
   box-sizing: border-box;
-  width: 100%;
   padding: 0.25rem;
   border-radius: 0.25rem;
-  display: flex;
+  display: inline-flex;
   flex-direction: column;
   justify-content: space-between;
+  gap: 0.25rem;
 
   border-style: solid;
   border-width: 1px;
@@ -19,115 +21,96 @@
 
   background-color: rgba(255, 255, 255, 0.8);
   box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.15);
+  cursor: pointer;
+
+  &.active {
+    border-color: var(--highlight-color);
+  }
 
   [data-theme="dark"] & {
     background-color: rgba(255, 255, 255, 0.15);
     box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.75);
   }
-}
 
-.active {
-  border-color: var(--highlight-color);
-}
+  .label {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.125rem;
+    margin: 0;
+    padding: 0;
+    font-size: 0.75rem;
+    line-height: 0.75rem;
+    height: 0.75rem;
+    transition: color 200ms;
+    user-select: none;
+    color: gray;
 
-.header {
-  box-sizing: border-box;
-  height: 0.75rem;
-  padding: 0 0.25rem;
-  display: flex;
-  justify-content: space-between;
-}
-
-.label {
-  display: inline;
-  margin: 0;
-  padding: 0;
-  font-size: 0.75rem;
-  line-height: 0.75rem;
-  height: 0.75rem;
-  vertical-align: top;
-  transition: color 200ms;
-  user-select: none;
-  color: var(--highlight-color);
-}
-
-.body {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-right: 0.25rem;
-  padding-left: 0.25rem;
-  cursor: pointer;
-}
-
-.select {
-  padding: 0.5rem;
-  width: 100%;
-  height: 1.25rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.selected {
-  width: 100%;
-  overflow: hidden;
-}
-
-.fallback {
-  opacity: 0.6;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.pulldown {
-  position: absolute;
-  top: 4rem;
-  left: 0;
-  z-index: 100;
-
-  box-sizing: border-box;
-  width: 100%;
-  padding: 0;
-  border-radius: 0.25rem;
-
-  background-color: rgba(255, 255, 255, 0.95);
-  box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.15);
-
-  [data-theme="dark"] & {
-    background-color: rgba(51, 51, 51, 0.95);
-    box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.75);
-  }
-}
-
-.option {
-  overflow: hidden;
-  width: 100%;
-  padding: 0.5rem;
-  box-sizing: border-box;
-  border-radius: 0.25rem;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 0.25rem;
-  transition: background-color 100ms;
-  cursor: pointer;
-
-  &:hover {
-    background-color: rgba(190, 194, 202, 0.3);
-    [data-theme="dark"] & {
-      background-color: rgba(190, 194, 202, 0.2);
+    &.label-active {
+      color: var(--highlight-color);
     }
   }
-}
 
-.description {
-  box-sizing: border-box;
-  opacity: 0.35;
-  padding-left: 0.25rem;
-  font-size: 0.75em;
-  white-space: nowrap;
-  text-overflow: ellipsis;
+  .body {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 0.25rem;
+    height: 1.25rem;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    .selected-option {
+      overflow: hidden;
+    }
+
+    .fallback {
+      opacity: 0.6;
+      display: flex;
+      justify-content: flex-start;
+      align-items: center;
+      gap: 0.5rem;
+    }
+
+    .pulldown {
+      position: absolute;
+      top: calc(100% + 0.25rem);
+      left: 0;
+      z-index: 100;
+
+      box-sizing: border-box;
+      width: max-content;
+      padding: 0;
+      border-radius: 0.25rem;
+
+      background-color: rgba(255, 255, 255, 0.95);
+      box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.15);
+
+      [data-theme="dark"] & {
+        background-color: rgba(51, 51, 51, 0.95);
+        box-shadow: 0 0 0.125rem rgba(0, 0, 0, 0.75);
+      }
+
+      .option {
+        overflow: hidden;
+        width: 100%;
+        padding: 0.5rem;
+        box-sizing: border-box;
+        border-radius: 0.25rem;
+        display: flex;
+        justify-content: flex-start;
+        align-items: center;
+        gap: 0.25rem;
+        transition: background-color 100ms;
+        cursor: pointer;
+
+        &:hover {
+          background-color: rgba(190, 194, 202, 0.3);
+          [data-theme="dark"] & {
+            background-color: rgba(190, 194, 202, 0.2);
+          }
+        }
+      }
+    }
+  }
 }
