+++ packages/qwik/src/components/form/elm-select.tsx
@@ -3,9 +3,10 @@ import {
   component$,
   PropsOf,
   useComputed$,
-  useOnDocument,
   useSignal,
+  useVisibleTask$,
   type CSSProperties,
+  type JSXOutput,
   type PropFunction,
 } from "@builder.io/qwik";
 import {
@@ -22,18 +23,43 @@ import { ElmCollapse } from "../containments/elm-collapse";
 
 export interface ElmSelectOption {
   id: string;
-  label: string;
-  description?: string;
+  slot: JSXOutput;
 }
 
 export interface ElmSelectProps extends PropsOf<"div"> {
+  /**
+   * Label for the select component.
+   */
   label: string;
-  placeholder?: string;
-  disabled?: boolean;
-  loading?: boolean;
-  options: ElmSelectOption[];
 
-  variant?: "default" | "small";
+  /**
+   * Icon for the select component.
+   * This is displayed on the left side of the label.
+   * If not provided, a default dropdown icon is used.
+   *
+   * @example <ElmSelectSlot icon={<ElmInlineIcon src={url} />} />
+   */
+  icon?: JSXOutput;
+
+  /**
+   * Placeholder text shown when no option is selected.
+   */
+  placeholder?: string;
+
+  /**
+   * Whether the select is disabled.
+   */
+  disabled?: boolean;
+
+  /**
+   * Whether the select is in a loading state.
+   */
+  loading?: boolean;
+
+  /**
+   * Options to display in the dropdown.
+   */
+  options: ElmSelectOption[];
 
   /**
    * Controlled selected option. When provided the parent owns the state.
@@ -78,6 +104,7 @@ export const ElmSelect = component$<ElmSelectProps>((props) => {
     disabled,
     loading,
     options,
+    icon,
     selectedOption: _selectedOptionProp,
     defaultSelectedOption,
     onSelectedOptionChange$,
@@ -101,27 +128,29 @@ export const ElmSelect = component$<ElmSelectProps>((props) => {
 
   const containerRef = useSignal<Element>();
 
-  useOnDocument(
-    "click",
-    $((event) => {
-      if (isOpen.value && containerRef.value) {
-        const target = event.target as Node;
-        if (!containerRef.value.contains(target)) {
-          setIsOpen(false);
-        }
+  // eslint-disable-next-line qwik/no-use-visible-task
+  useVisibleTask$(({ cleanup }) => {
+    const handler = (event: MouseEvent) => {
+      if (!isOpen.value || !containerRef.value) return;
+      const target = event.target as Node;
+      if (!containerRef.value.contains(target)) {
+        void setIsOpen(false);
       }
-    }),
-  );
+    };
+    document.addEventListener("click", handler);
+    cleanup(() => document.removeEventListener("click", handler));
+  });
 
   return (
     <div
       ref={containerRef}
       class={[styles.wrapper, isOpen.value && styles.active, className]}
-      style={{
-        backgroundColor: disabled || loading ? "rgba(0,0,0,0.15)" : undefined,
-        "--highlight-color": isOpen.value ? "#bfa056" : undefined,
-        ...(style as CSSProperties),
-      }}
+      style={
+        {
+          backgroundColor: disabled || loading ? "rgba(0,0,0,0.15)" : undefined,
+          ...(style as CSSProperties),
+        } as CSSProperties
+      }
       onClick$={$(() => {
         if (!props.disabled && !props.loading) {
           setIsOpen(!isOpen.value);
@@ -129,52 +158,44 @@ export const ElmSelect = component$<ElmSelectProps>((props) => {
       })}
       {...rest}
     >
-      <div class={styles.header}>
-        <span class={[styles.label, textStyles.text]}>{label}</span>
-      </div>
+      <span class={[styles.label, { [styles["label-active"]]: isOpen.value }]}>
+        {icon ? (
+          <div class={styles.icon}>{icon}</div>
+        ) : (
+          <ElmMdiIcon d={mdiArrowDownDropCircleOutline} size="0.75rem" />
+        )}
+        {label}
+      </span>
 
       <div class={styles.body}>
-        <div class={styles.select}>
-          <div class={[styles.selected, textStyles.text]}>
-            {selectedOption.value ? (
-              <div key={selectedOption.value.id}>
-                <span>{selectedOption.value.label}</span>
-                {selectedOption.value.description && (
-                  <span class={styles.description}>
-                    {selectedOption.value.description}
-                  </span>
-                )}
-              </div>
-            ) : (
-              <div class={styles.fallback}>
-                <ElmMdiIcon d={mdiArrowDownDropCircleOutline} />
-                <span>{placeholder ?? "Select an option"}</span>
-              </div>
-            )}
-          </div>
-
-          <ElmMdiIcon d={mdiMenuDown} size="1.5rem" />
-
-          <ElmCollapse isOpen={isOpen.value} class={styles.pulldown}>
-            {options.map((option) => (
-              <div
-                key={option.id}
-                class={[styles.option, textStyles.text]}
-                onClick$={(e) => {
-                  e.stopPropagation();
-                  setSelectedOption(option);
-                  setIsOpen(false);
-                }}
-              >
-                <ElmMdiIcon d={mdiChevronRight} color="#868e9c" size="0.75em" />
-                <span>{option.label}</span>
-                {option.description && (
-                  <span class={styles.description}>{option.description}</span>
-                )}
-              </div>
-            ))}
-          </ElmCollapse>
+        <div class={[styles["selected-option"], textStyles.text]}>
+          {selectedOption.value ? (
+            <div key={selectedOption.value.id}>{selectedOption.value.slot}</div>
+          ) : (
+            <div class={styles.fallback}>
+              <span>{placeholder ?? "Select an option"}</span>
+            </div>
+          )}
         </div>
+
+        <ElmMdiIcon d={mdiMenuDown} size="1.5rem" />
+
+        <ElmCollapse isOpen={isOpen.value} class={styles.pulldown}>
+          {options.map((option) => (
+            <div
+              key={option.id}
+              class={[styles.option, textStyles.text]}
+              onClick$={(e) => {
+                e.stopPropagation();
+                setSelectedOption(option);
+                setIsOpen(false);
+              }}
+            >
+              <ElmMdiIcon d={mdiChevronRight} color="#868e9c" size="0.75em" />
+              {option.slot}
+            </div>
+          ))}
+        </ElmCollapse>
       </div>
     </div>
   );
